feat #76: home-api-connect - admin response 수정
