+++ app/src/main/java/com/example/assu_fe_app/data/dto/partnership/response/GetProposalPartnerListResponseDto.kt
@@ -1,37 +1,28 @@
 package com.example.assu_fe_app.data.dto.partnership.response
 
 import com.example.assu_fe_app.domain.model.admin.GetProposalPartnerListModel
+import com.example.assu_fe_app.domain.model.admin.PartnershipGoodsModel
+import com.example.assu_fe_app.domain.model.admin.PartnershipOptionModel
 import com.example.assu_fe_app.domain.model.admin.ProposalPartnerDetailsModel
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class GetProposalPartnerListResponseDto(
-    val shopName: String,
-    val content: List<ProposalPartnerDetails>,
-    val startDate: String,
-    val endDate: String,
+    val partnershipId: Long,
+    val partnershipPeriodStart: String,
+    val partnershipPeriodEnd: String,
+    val adminId: Long,
+    val partnerId: Long,
+    val storeId: Long,
+    val options: List<PartnershipOption>
 ) {
     fun toModel() = GetProposalPartnerListModel (
-        shopName = this.shopName,
-        content = this.content.map{it.toModel()},
-        startDate = this.startDate,
-        endDate = this.endDate,
-    )
-}
-
-@JsonClass(generateAdapter = true)
-data class ProposalPartnerDetails (
-    val type: String,
-    val people: Long,
-    val cost: Long?,      // nullable 처리
-    val discount: Long?,  // nullable 처리
-    val goods: List<String>
-) {
-    fun toModel() = ProposalPartnerDetailsModel (
-        type = this.type,
-        people = this.people,
-        cost = this.cost,
-        discount = this.discount,
-        goods = this.goods
+        partnershipId = this.partnershipId,
+        partnershipPeriodStart = org.threeten.bp.LocalDate.parse(partnershipPeriodStart),
+        partnershipPeriodEnd = org.threeten.bp.LocalDate.parse(partnershipPeriodEnd),
+        adminId = this.adminId,
+        partnerId = this.partnerId,
+        storeId = this.storeId,
+        options = this.options.map { it.toModel() }
     )
 }
