+++ app/src/main/java/com/example/assu_fe_app/domain/model/admin/GetProposalPartnerListModel.kt
@@ -1,8 +1,11 @@
 package com.example.assu_fe_app.domain.model.admin
 
 data class GetProposalPartnerListModel(
-    val shopName: String,
-    val content: List<ProposalPartnerDetailsModel>,
-    val startDate: String,
-    val endDate: String,
+    val partnershipId: Long,
+    val partnershipPeriodStart: org.threeten.bp.LocalDate,
+    val partnershipPeriodEnd: org.threeten.bp.LocalDate,
+    val adminId: Long,
+    val partnerId: Long,
+    val storeId: Long,
+    val options: List<PartnershipOptionModel>
 )
