+++ app/src/main/java/com/example/assu_fe_app/presentation/admin/home/AdminHomeFragment.kt
@@ -93,84 +93,6 @@ class AdminHomeFragment :
             }
         }
 
-
-        viewLifecycleOwner.lifecycleScope.launch {
-            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                partnershipViewModel.getPartnershipPartnerListUiState.collect { state ->
-                    when (state) {
-                        is PartnershipViewModel.PartnershipPartnerListUiState.Success -> {
-                            val list = state.data
-
-                            if(list.isEmpty()) {
-                                binding.btnAdminHomeViewAll.visibility = View.INVISIBLE
-                                binding.tvNoPartnerList.visibility = View.VISIBLE
-                            } else {
-                                binding.tvNoPartnerList.visibility = View.GONE
-                            }
-
-                            // 1번째 카드
-                            if (list.isNotEmpty()) {
-                                val item1 = list[0]
-                                binding.tvPartnerName1.text = item1.shopName
-                                binding.tvBenefitDescription1.text =
-                                    item1.content.firstOrNull()?.type ?: "제휴 혜택 없음"
-                                binding.tvBenefitPeriod1.text =
-                                    "${item1.startDate} ~ ${item1.endDate}"
-                                binding.adminHomeListItem1.visibility = View.VISIBLE
-
-                                binding.adminHomeListItem1.setOnClickListener {
-                                    val dialog = PartnershipContractDialogFragment(
-                                        item1.content.map { c ->
-                                            PartnershipContractItem.Service.ByPeople(
-                                                c.people.toInt(), // UI 모델이 Int면 toInt(), Long으로 바꾸면 그대로 사용
-                                                c.type
-                                            )
-                                        }
-                                    )
-                                    dialog.show(parentFragmentManager, "PartnershipContentDialog")
-                                }
-                            } else {
-                                binding.adminHomeListItem1.visibility = View.GONE
-                            }
-
-                            // 2번째 카드
-                            if (list.size > 1) {
-                                val item2 = list[1]
-                                binding.tvAdminName2.text = item2.shopName
-                                binding.tvBenefitDescription2.text =
-                                    item2.content.firstOrNull()?.type ?: "제휴 혜택 없음"
-                                binding.tvBenefitPeriod2.text =
-                                    "${item2.startDate} ~ ${item2.endDate}"
-                                binding.adminHomeListItem2.visibility = View.VISIBLE
-
-                                binding.adminHomeListItem2.setOnClickListener {
-                                    val dialog = PartnershipContractDialogFragment(
-                                        item2.content.map { c ->
-                                            PartnershipContractItem.Service.ByPeople(
-                                                c.people.toInt(),
-                                                c.type
-                                            )
-                                        }
-                                    )
-                                    dialog.show(parentFragmentManager, "PartnershipContentDialog")
-                                }
-                            } else {
-                                binding.adminHomeListItem2.visibility = View.GONE
-                            }
-                        }
-
-                        // ✅ 나머지 모든 상태에선 가려두기
-                        is PartnershipViewModel.PartnershipPartnerListUiState.Loading,
-                        is PartnershipViewModel.PartnershipPartnerListUiState.Idle,
-                        is PartnershipViewModel.PartnershipPartnerListUiState.Fail,
-                        is PartnershipViewModel.PartnershipPartnerListUiState.Error -> {
-                            binding.adminHomeListItem1.visibility = View.GONE
-                            binding.adminHomeListItem2.visibility = View.GONE
-                        }
-                    }
-                }
-            }
-        }
     }
 
     override fun onResume() {
